feat: iNaturalist api for searching; fix: ui, history saving on web; refactor: move llm api logic from viewmodel

+++ src/composeApp/src/androidMain/kotlin/hcmus/bugscanner/core/utils/TimeUtils.android.kt
@@ -7,12 +7,12 @@ import java.util.Locale
 /**
  * Khai báo hàm yêu cầu Android cung cấp cách lấy thời gian
  */
-actual fun getCurrentTimeMillis(): Long = System.currentTimeMillis()
+actual fun getCurrentTimeMillis(): Double = System.currentTimeMillis().toDouble()
 
 /**
  * Khai báo hàm yêu cầu Android cung cấp cách format thời gian
  */
-actual fun formatTimestamp(timestamp: Long): String {
+actual fun formatTimestamp(timestamp: Double): String {
     val sdf = SimpleDateFormat("dd/MM/yyyy HH:mm", Locale.getDefault())
-    return sdf.format(Date(timestamp))
+    return sdf.format(Date(timestamp.toLong()))
 }
