+++ src/composeApp/src/commonMain/kotlin/hcmus/bugscanner/core/utils/TimeUtils.kt
@@ -3,9 +3,9 @@ package hcmus.bugscanner.core.utils
 /**
  * Khai báo hàm yêu cầu cung cấp cách lấy thời gian
  */
-expect fun getCurrentTimeMillis(): Long
+expect fun getCurrentTimeMillis(): Double
 
 /**
  * Khai báo hàm yêu cầu cung cấp cách format thời gian
  */
-expect fun formatTimestamp(timestamp: Long): String
+expect fun formatTimestamp(timestamp: Double): String
